# Server side: - Removed TaskComment entity / repository / service / controller - Implemented pagination to Task module - Removed sorting from Task, replaced with sort in TaskService

+++ server/src/main/java/pl/readyTask/service/TaskService.java
@@ -1,34 +1,62 @@
 package pl.readyTask.service;
 
 import lombok.AllArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import pl.readyTask.entity.Task;
+import pl.readyTask.entity.Team;
 import pl.readyTask.entity.User;
+import pl.readyTask.entity.enumeration.MemberRole;
 import pl.readyTask.entity.enumeration.TaskState;
 import pl.readyTask.exception.AccessDeniedToActionException;
 import pl.readyTask.exception.NoDataFoundException;
 import pl.readyTask.repository.TaskRepository;
 
-import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
 public class TaskService {
     private final TaskRepository taskRepository;
     private final SecurityService securityService;
-    private static final Logger logger = LoggerFactory.getLogger(TaskService.class);
+    private final MembershipService membershipService;
     private final List<TaskState> statesAllowedToSetForUserAssignedToTask = List.of(TaskState.IN_PROGRESS, TaskState.TO_REVIEW);
     private final List<TaskState> statesAllowedToSetForAuthorOfTask = List.of(TaskState.TO_FIX, TaskState.FINISHED, TaskState.ARCHIVED);
 
     public Task getById(UUID id) {
         return taskRepository.findById(id).orElseThrow(() -> new NoDataFoundException("task", id));
+    }
+
+    public Page<Task> getPagedByUserIdAndTeamAdminRole(Long userId, int page) {
+        Pageable pageable = getPageable(page);
+        return taskRepository.findAllTasksManagedByUser(userId, MemberRole.ADMIN, pageable);
+    }
+
+    public Page<Task> getPagedByTeamId(Long teamId, int page) {
+        Pageable pageable = getPageable(page);
+        return taskRepository.findAllByTeamId(teamId, pageable);
+    }
+
+    public Page<Task> getPagedByUserAssignedToAndTeamId(Long userId, Long teamId, int page) {
+        Pageable pageable = getPageable(page);
+        return taskRepository.findByUserAssignedToTaskIdAndTeamId(userId, teamId, pageable);
+    }
+
+    public Page<Task> getPagedByUserAssignedToId(Authentication authentication, int page) {
+        User user = securityService.getUserByEmailFromAuthentication(authentication);
+        Pageable pageable = getPageable(page);
+        return taskRepository.findByUserAssignedToTaskIdOrderByState(user.getId(), pageable);
+    }
+
+    private Pageable getPageable(int page) {
+        int pageSize = 6;
+        return PageRequest.of(page, pageSize, Sort.by("state").and(Sort.by("deadline")));
     }
 
     public Task add(Task task, Authentication authentication) {
@@ -38,59 +66,54 @@ public class TaskService {
         return taskRepository.save(task);
     }
 
-    public List<Task> getByUserAssignedToId(Authentication authentication) {
-        User user = securityService.getUserByEmailFromAuthentication(authentication);
-        return taskRepository.findByUserAssignedToTaskIdOrderByState(user.getId())
-                .orElseThrow(() -> new NoDataFoundException(Task.class.getName(), "userId", user.getId()));
-    }
-
     public void changeTaskState(UUID taskId, TaskState state, Authentication authentication) {
         User user = this.securityService.getUserByEmailFromAuthentication(authentication);
         Task task = this.taskRepository.findById(taskId)
                 .orElseThrow(() -> new NoDataFoundException(Task.class.getName(), taskId));
+        checkIsUserAllowedToChangeTaskState(user, task, state);
+        task.setState(state);
+        taskRepository.save(task);
+    }
 
+    private void checkIsUserAllowedToChangeTaskState(User user, Task task, TaskState state) {
         boolean isAuthorOfTaskAllowedToChangeState =
                 isUserAuthorOfTask(user, task) && statesAllowedToSetForAuthorOfTask.contains(state);
         boolean isUserAssignedToTaskAllowedToChangeState =
                 isUserAssignedToTask(user, task) && statesAllowedToSetForUserAssignedToTask.contains(state);
-
-        if(isAuthorOfTaskAllowedToChangeState || isUserAssignedToTaskAllowedToChangeState){
-            task.setState(state);
-            taskRepository.save(task);
-        }else{
+        if (!(isAuthorOfTaskAllowedToChangeState || isUserAssignedToTaskAllowedToChangeState)) {
             throw new AccessDeniedToActionException(AccessDeniedToActionException.changeTaskStateMessage);
         }
-    }
-
-    private boolean isUserAuthorOfTask(User user, Task task){
-        return Objects.equals(user.getId(), task.getAuthorOfTask().getId());
-    }
-
-    private void checkIsUserRelatedToTask(boolean isRelated){
-        if(!isRelated){
-            throw new AccessDeniedToActionException(AccessDeniedToActionException.noPermissionToTask);
-        }
-    }
-
-    private boolean isUserAssignedToTask(User user, Task task){
-        return Objects.equals(user.getId(), task.getUserAssignedToTask().getId());
     }
 
     public void deleteById(UUID taskId, Authentication authentication) {
         User user = this.securityService.getUserByEmailFromAuthentication(authentication);
         Task task = this.taskRepository.findById(taskId)
                 .orElseThrow(() -> new NoDataFoundException(Task.class.getName(), taskId));
-        boolean isUserAuthorOfTask = isUserAuthorOfTask(user, task);
-        checkIsUserRelatedToTask(isUserAuthorOfTask);
+        checkIsUserAdminOfTeamRelatedToTask(user, task.getTeam());
         taskRepository.deleteById(taskId);
     }
 
-    public List<Task> getByUserAssignedToAndTeamId(Long userId, Long teamId) {
-        return taskRepository.findByUserAssignedToTaskIdAndTeamId(userId, teamId)
-                .orElseThrow(() -> new NoDataFoundException(
-                        Task.class.getName(), "userAssignedToTaskId, teamId", userId + ' ' + teamId))
-                .stream()
-                .sorted(Task.getComparator())
-                .collect(Collectors.toList());
+    private boolean isUserAuthorOfTask(User user, Task task) {
+        return Objects.equals(user.getId(), task.getAuthorOfTask().getId());
+    }
+
+    private boolean isUserAssignedToTask(User user, Task task) {
+        return Objects.equals(user.getId(), task.getUserAssignedToTask().getId());
+    }
+
+    private void checkIsUserRelatedToTask(boolean isRelated) {
+        if (!isRelated) {
+            throw new AccessDeniedToActionException(AccessDeniedToActionException.noPermissionToTask);
+        }
+    }
+
+    public void checkIsUserAllowedToGetTask(User user, Task task) {
+        boolean isUserMemberOfTeam = membershipService.isUserMemberOfTeam(user, task.getTeam());
+        boolean isUserAdminOfTeamRelatedToTask = membershipService.isUserAdminOfTeam(user, task.getTeam());
+        checkIsUserRelatedToTask(isUserMemberOfTeam || isUserAdminOfTeamRelatedToTask);
+    }
+
+    public void checkIsUserAdminOfTeamRelatedToTask(User user, Team team) {
+        membershipService.isUserAdminOfTeam(user, team);
     }
 }
